pom和flyway

+++ src/main/resources/db/migration/V1__Create_user.sql
@@ -1,6 +1,6 @@
 create table user
 (
-    id                 int primary key AUTO_INCREMENT,
+    id                 bigint primary key AUTO_INCREMENT,
     username           varchar(10) unique,
     encrypted_password varchar(100),
     avatar             varchar(100),
